+++ src/Beef.AspNetCore.WebApi/WebApiStartup.cs
@@ -1,7 +1,14 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/Beef
 
 using Beef.Diagnostics;
+using Microsoft.AspNetCore;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Azure.KeyVault;
+using Microsoft.Azure.Services.AppAuthentication;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Configuration.AzureKeyVault;
+using Microsoft.Extensions.FileProviders;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using System;
 
@@ -12,6 +19,60 @@ namespace Beef.AspNetCore.WebApi
     /// </summary>
     public static class WebApiStartup
     {
+        /// <summary>
+        /// Builds (creates) the <see cref="IWebHost"/> using the <see cref="WebHost.CreateDefaultBuilder(string[])"/> utilizing the <see cref="ConfigurationBuilder"/>.
+        /// </summary>
+        /// <typeparam name="TStartup">The API startup <see cref="Type"/>.</typeparam>
+        /// <param name="args">The command line args.</param>
+        /// <param name="environmentVariablePrefix">The prefix that the environment variables must start with.</param>
+        /// <returns>The <see cref="IWebHost"/>.</returns>
+        public static IWebHost BuildWebHost<TStartup>(string[] args, string? environmentVariablePrefix = null) where TStartup : class =>
+            WebHost.CreateDefaultBuilder(args)
+                   .ConfigureAppConfiguration((hostingContext, config) => ConfigurationBuilder<TStartup>(config, hostingContext.HostingEnvironment, environmentVariablePrefix))
+                   .UseStartup<TStartup>()
+                   .Build();
+
+        /// <summary>
+        /// Builds the configuration probing; will probe in the following order: 1) Azure Key Vault (see https://docs.microsoft.com/en-us/aspnet/core/security/key-vault-configuration) 
+        /// or User Secrets where hosting environment is development (see https://docs.microsoft.com/en-us/aspnet/core/security/app-secrets), 2) environment variable (see <paramref name="environmentVariablePrefix"/>),
+        /// 3) appsettings.{environment}.json, 4) appsettings.json, 5) webapisettings.{environment}.json (embedded resource), and 6) webapisettings.json (embedded resource).
+        /// </summary>
+        /// <typeparam name="TStartup">The API startup <see cref="Type"/>.</typeparam>
+        /// <param name="configurationBuilder">The <see cref="IConfigurationBuilder"/>.</param>
+        /// <param name="hostingEnvironment">The <see cref="IWebHostEnvironment"/>.</param>
+        /// <param name="environmentVariablePrefix">The prefix that the environment variables must start with.</param>
+        public static void ConfigurationBuilder<TStartup>(IConfigurationBuilder configurationBuilder, IWebHostEnvironment hostingEnvironment, string? environmentVariablePrefix = null) where TStartup : class
+        {
+            if (configurationBuilder == null)
+                throw new ArgumentNullException(nameof(configurationBuilder));
+
+            if (hostingEnvironment == null)
+                throw new ArgumentNullException(nameof(hostingEnvironment));
+
+            configurationBuilder.AddJsonFile(new EmbeddedFileProvider(typeof(TStartup).Assembly), $"webapisettings.json", true, false)
+                .AddJsonFile(new EmbeddedFileProvider(typeof(TStartup).Assembly), $"webapisettings.{hostingEnvironment.EnvironmentName}.json", true, false)
+                .AddJsonFile("appsettings.json", true, true)
+                .AddJsonFile($"appsettings.{hostingEnvironment.EnvironmentName}.json", true, true)
+                .AddEnvironmentVariables(environmentVariablePrefix);
+
+            var config = configurationBuilder.Build();
+            if (hostingEnvironment.IsDevelopment())
+            {
+                if (config.GetValue<bool>("UseUserSecrets"))
+                    configurationBuilder.AddUserSecrets<TStartup>();
+            }
+            else
+            {
+                var kvn = config["KeyVaultName"];
+                if (!string.IsNullOrEmpty(kvn))
+                {
+                    var astp = new AzureServiceTokenProvider();
+                    using var kvc = new KeyVaultClient(new KeyVaultClient.AuthenticationCallback(astp.KeyVaultTokenCallback));
+                    configurationBuilder.AddAzureKeyVault($"https://{kvn}.vault.azure.net/", kvc, new DefaultKeyVaultSecretManager());
+                }
+            }
+        }
+
         /// <summary>
         /// Binds (redirects) Beef <see cref="Beef.Diagnostics.Logger"/> to the ASP.NET Core <see cref="Microsoft.Extensions.Logging.ILogger"/>.
         /// </summary>